Added progress tracking
Progress tracking is accomplished by setting a cookie
named 'pageless_progress' which stores the element ID
of the 'current' display element.

Tracking the current element requires defining a selector
for the elements you'd like to track and passing this
selector to pageless as the 'progress' parameter.

When the page loads the first element matching the selector
will be set as the current element.  Once the top of
the current element is scrolled above the top of the window
the next matching element will be set to current.